+++ src/GlobalPayments.Api/Entities/Enums/BNPLShippingMethod.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GlobalPayments.Api.Entities
+{
+    public enum BNPLShippingMethod
+    {
+         DELIVERY,
+         COLLECTION,
+         EMAIL,
+    }
+}