+++ src/GlobalPayments.Api/Entities/GpApi/GpApiAuthorizationRequestBuilder.cs
@@ -1,6 +1,7 @@
 ﻿using GlobalPayments.Api.Builders;
 using GlobalPayments.Api.Gateways;
 using GlobalPayments.Api.PaymentMethods;
+using GlobalPayments.Api.PaymentMethods.PaymentInterfaces;
 using GlobalPayments.Api.Utils;
 using System;
 using System.Collections.Generic;
@@ -241,6 +242,8 @@ namespace GlobalPayments.Api.Entities {
                 if (secureEcom != null)
                 {
                     var authentication = new JsonDoc().Set("id", secureEcom.ServerTransactionId);
+                    var three_ds =new JsonDoc().Set("exempt_status", secureEcom.ExemptStatus.ToString());
+                    authentication.Set("three_ds", three_ds);
 
                     paymentMethod.Set("authentication", authentication);
                 }
@@ -296,6 +299,15 @@ namespace GlobalPayments.Api.Entities {
                     .Set("provider", alternatepaymentMethod.AlternativePaymentMethodType?.ToString()?.ToLower())
                     .Set("address_override_mode", alternatepaymentMethod.AddressOverrideMode);
                 paymentMethod.Set("apm", apm);
+            }
+
+            if (builder.PaymentMethod is BNPL) {
+                BNPL bnpl = (BNPL)builder.PaymentMethod;
+
+                var bnplType = new JsonDoc().Set("provider", EnumConverter.GetMapping(Target.GP_API, bnpl.BNPLType));
+
+                paymentMethod.Set("name", builder.CustomerData?.FirstName + " " + builder.CustomerData?.LastName)
+                    .Set("bnpl", bnplType);
             }
 
             // encryption
@@ -393,7 +405,7 @@ namespace GlobalPayments.Api.Entities {
                 .Set("link", !string.IsNullOrEmpty(builder.PaymentLinkId) ? new JsonDoc()
                 .Set("id", builder.PaymentLinkId) : null);
 
-            if (builder.PaymentMethod is eCheck || builder.PaymentMethod is AlternativePaymentMethod) {
+            if (builder.PaymentMethod is eCheck || builder.PaymentMethod is AlternativePaymentMethod || builder.PaymentMethod is BNPL) {
                 data.Set("payer", SetPayerInformation(builder));
             }
 
@@ -406,15 +418,22 @@ namespace GlobalPayments.Api.Entities {
                 data.Set("order", order);
             }
 
-            if (builder.PaymentMethod is AlternativePaymentMethod) {
-                setOrderInformation(builder, ref data);
+            if (builder.PaymentMethod is AlternativePaymentMethod || builder.PaymentMethod is BNPL) {
+                SetOrderInformation(builder, ref data);
 
-                var alternatepaymentMethod = (AlternativePaymentMethod)builder.PaymentMethod;
+                INotificationData payment = null;
+
+                if (builder.PaymentMethod is AlternativePaymentMethod) {
+                    payment = (builder.PaymentMethod) as AlternativePaymentMethod;                   
+                }
+                if (builder.PaymentMethod is BNPL) {
+                    payment = builder.PaymentMethod as BNPL;                   
+                }
 
                 var notifications = new JsonDoc()
-                   .Set("return_url", alternatepaymentMethod?.ReturnUrl)
-                   .Set("status_url", alternatepaymentMethod?.StatusUpdateUrl)
-                   .Set("cancel_url", alternatepaymentMethod?.CancelUrl);
+                      .Set("return_url", payment?.ReturnUrl)
+                      .Set("status_url", payment?.StatusUpdateUrl)
+                      .Set("cancel_url", payment?.CancelUrl);
 
                 data.Set("notifications", notifications);
             }
@@ -576,6 +595,52 @@ namespace GlobalPayments.Api.Entities {
 
                 payer.Set("work_phone", workPhone);
             }
+            else if(builder.PaymentMethod is BNPL && builder.CustomerData != null) {
+               
+                payer.Set("email", builder.CustomerData.Email)
+                     .Set("date_of_birth", builder.CustomerData.DateOfBirth);
+
+                JsonDoc billing_address = new JsonDoc();
+
+                billing_address.Set("line_1", builder.BillingAddress?.StreetAddress1)
+                            .Set("line_2", builder.BillingAddress?.StreetAddress2)
+                            .Set("city", builder.BillingAddress?.City)
+                            .Set("postal_code", builder.BillingAddress?.PostalCode)
+                            .Set("state", builder.BillingAddress?.State)
+                            .Set("country", builder.BillingAddress?.CountryCode);
+                
+                if (builder.CustomerData != null) {
+                    billing_address
+                           .Set("first_name", builder.CustomerData?.FirstName)
+                           .Set("last_name", builder.CustomerData?.LastName);
+                }
+
+                payer.Set("billing_address", billing_address);
+
+                if (builder.CustomerData.Phone != null) {
+                    JsonDoc homePhone = new JsonDoc();
+
+                    homePhone.Set("country_code", builder.CustomerData.Phone?.CountryCode)
+                            .Set("subscriber_number", builder.CustomerData.Phone?.Number);
+
+                    payer.Set("contact_phone", homePhone);
+                    
+                }
+                if (builder.CustomerData.Documents != null) {
+
+                    var documents = new List<Dictionary<string, object>>();
+                    foreach (var document in builder.CustomerData.Documents) {
+                        var doc = new Dictionary<string, object>();
+                        
+                        doc.Add("type", document.Type.ToString());
+                        doc.Add("reference", document.Reference);
+                        doc.Add("issuer", document.Issuer);
+
+                        documents.Add(doc);                            
+                    }
+                    payer.Set("documents", documents);
+                }
+            }
             return payer;
         }
 
@@ -589,47 +654,41 @@ namespace GlobalPayments.Api.Entities {
             return "AUTO";
         }
 
-        private static JsonDoc setOrderInformation(AuthorizationBuilder builder, ref JsonDoc requestBody)
+        private static JsonDoc SetItemDetailsListForBNPL(AuthorizationBuilder builder, ref JsonDoc order)
         {
-            JsonDoc order;
-            if (requestBody.Has("order"))
-            {
-                order = requestBody.Get("order");
-            }
-            else
-            {
-                order = new JsonDoc();
-            }
-            order.Set("description", builder.OrderDetails?.Description);
+            var items = new List<Dictionary<string, object>>();                        
+                foreach (var product in builder.MiscProductData) {
+                    var item = new Dictionary<string, object>();                    
+                    var qta = product.Quantity;
+                    var taxAmount = product.TaxAmount ?? 0;
+                    var unitAmount = product.UnitPrice ?? 0;
+                    var netUnitAmount = product.NetUnitAmount ?? 0;
+                    var discountAmount = product.DiscountAmount ?? 0;
+                    item.Add("reference", product.ProductId ?? null);
+                    item.Add("label", product.ProductName ?? null);
+                    item.Add("description", product.Description ?? null);
+                    item.Add("quantity", qta.ToString());
+                    item.Add("unit_amount", unitAmount.ToNumericCurrencyString());
+                    item.Add("total_amount", (qta * unitAmount).ToNumericCurrencyString());
+                    item.Add("tax_amount", taxAmount.ToNumericCurrencyString());
+                    item.Add("discount_amount", discountAmount != 0 ? discountAmount.ToNumericCurrencyString().RemoveInitialZero() : "0");
+                    item.Add("tax_percentage", product.TaxPercentage != 0 ? product.TaxPercentage.ToNumericCurrencyString().RemoveInitialZero() : "0");
+                    item.Add("net_unit_amount", netUnitAmount.ToNumericCurrencyString());
+                    item.Add("gift_card_currency", product.GiftCardCurrency);
+                    item.Add("url", product.Url);
+                    item.Add("image_url", product.ImageUrl);
+                    items.Add(item);
+                }
+                         
+            return order.Set("items", items);
+        }
 
-            if (builder.ShippingAddress != null)
-            {
-                var shippingAddress = new JsonDoc()
-                    .Set("line1", builder.ShippingAddress.StreetAddress1)
-                    .Set("line2", builder.ShippingAddress.StreetAddress2)
-                    .Set("line3", builder.ShippingAddress.StreetAddress3)
-                    .Set("city", builder.ShippingAddress.City)
-                    .Set("postal_code", builder.ShippingAddress.PostalCode)
-                    .Set("state", builder.ShippingAddress.State)
-                    .Set("country", builder.ShippingAddress.CountryCode);
-
-                order.Set("shipping_address", shippingAddress);
-            }
-
-            var shippingPhone = new JsonDoc()
-                .Set("country_code", builder.ShippingPhone?.CountryCode)
-                .Set("subscriber_number", builder.ShippingPhone?.Number);
-
-            order.Set("shipping_phone", shippingPhone);
-
-            decimal taxTotalAmount = 0;
-            decimal itemsAmount = 0;
+        private static JsonDoc SetItemDetailsListForApm(AuthorizationBuilder builder, ref JsonDoc order) {
+            decimal taxTotalAmount = 0, itemsAmount = 0;
             decimal? orderAmount = null;
-            if (builder.MiscProductData != null)
-            {
-                var items = new List<Dictionary<string,object>>();
-                foreach (var product in builder.MiscProductData)
-                {
+            if (builder.MiscProductData != null) {
+                var items = new List<Dictionary<string, object>>();
+                foreach (var product in builder.MiscProductData) {
                     var qta = product.Quantity ?? 0;
                     var taxAmount = product.TaxAmount ?? 0;
                     var unitAmount = product.UnitPrice ?? 0;
@@ -662,8 +721,65 @@ namespace GlobalPayments.Api.Entities {
                 order.Set("currency", builder.Currency);
                 order.Set("items", items);
             }
-            if (!requestBody.Has("order"))
-            {
+
+            return order;
+        }
+
+        private static JsonDoc SetOrderInformation(AuthorizationBuilder builder, ref JsonDoc requestBody) {
+            JsonDoc order;
+            if (requestBody.Has("order")) {
+                order = requestBody.Get("order");
+            }
+            else {
+                order = new JsonDoc();
+            }
+            order.Set("description", builder.OrderDetails?.Description);
+
+            var shippingAddress = new JsonDoc();
+            if (builder.ShippingAddress != null) {
+                shippingAddress
+                    .Set("line_1", builder.ShippingAddress.StreetAddress1)
+                    .Set("line_2", builder.ShippingAddress.StreetAddress2)
+                    .Set("line_3", builder.ShippingAddress.StreetAddress3)
+                    .Set("city", builder.ShippingAddress.City)
+                    .Set("postal_code", builder.ShippingAddress.PostalCode)
+                    .Set("state", builder.ShippingAddress.State)
+                    .Set("country", builder.ShippingAddress.CountryCode);                
+            }
+
+            var shippingPhone = new JsonDoc()
+                .Set("country_code", builder.ShippingPhone?.CountryCode)
+                .Set("subscriber_number", builder.ShippingPhone?.Number);
+
+            if (shippingPhone.HasKeys()) {
+                order.Set("shipping_phone", shippingPhone);
+            }
+
+            //AlternativePaymentMethod
+            if (builder.PaymentMethod is AlternativePaymentMethod) { 
+                if (builder.MiscProductData != null) {
+                    SetItemDetailsListForApm(builder, ref order);
+                }
+            }
+
+            //Buy Now Pay Later
+            if (builder.PaymentMethod is BNPL) {
+                order.Set("shipping_method", builder.BNPLShippingMethod.ToString());
+
+                if (builder.MiscProductData != null) {
+                    SetItemDetailsListForBNPL(builder, ref order);
+                }
+
+                if (builder.CustomerData != null) {
+                     shippingAddress
+                            .Set("first_name", builder.CustomerData?.FirstName)
+                            .Set("last_name", builder.CustomerData?.LastName);                    
+                }
+            }
+
+            order.Set("shipping_address", shippingAddress);
+
+            if (!requestBody.Has("order")) {
                 requestBody.Set("order", order);
             }