+++ src/GlobalPayments.Api/Entities/Product.cs
@@ -10,6 +10,7 @@ namespace GlobalPayments.Api.Entities {
         public string Description { get; set; }
         public int? Quantity { get; set; }
         public decimal? UnitPrice { get; set; }
+        public decimal? NetUnitPrice { get; set; }
         public bool? Gift { get; set; }
         public string UnitCurrency { get; set; }
         public string Type { get; set; }
@@ -19,5 +20,7 @@ namespace GlobalPayments.Api.Entities {
         public decimal? DiscountAmount { get; set; }
         public string Url { get; set; }
         public string ImageUrl { get; set; }
+        public decimal? NetUnitAmount { get; set; }
+        public string GiftCardCurrency { get; set; }
     }
 }