+++ src/GlobalPayments.Api/PaymentMethods/PaymentInterfaces/IPaymentMethod.cs
@@ -2,6 +2,6 @@
 
 namespace GlobalPayments.Api.PaymentMethods {
     public interface IPaymentMethod {
-        PaymentMethodType PaymentMethodType { get; }
+        PaymentMethodType PaymentMethodType { get; } 
     }
 }